+++ packages/axis-configuration/src/shared/errors/UnauthorizedError.ts
@@ -1,4 +1,4 @@
 /**
  * Error thrown when user is unauthorized to perform a certain operation.
  */
-export class UnauthorizationError extends Error {}
+export class UnauthorizedError extends Error {}