+++ packages/axis-configuration/test/shared/Connection.spec.ts
@@ -6,32 +6,32 @@ describe('connection', () => {
     describe('#ctor(protocol, ...)', () => {
         test('should return connection without options', () => {
             // Act
-            const connection = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass');
+            const got = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass');
 
             // Assert
-            expect(connection).toBeTruthy();
+            expect(got).toBeTruthy();
         });
 
         test('should return connection with http agent options', () => {
             // Act
-            const connection = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
+            const got = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
                 agent: new http.Agent(),
             });
 
             // Assert
-            expect(connection).toBeTruthy();
-            expect(connection.options?.agent).toBeTruthy();
+            expect(got).toBeTruthy();
+            expect(got.options?.agent).toBeTruthy();
         });
 
         test('should return connection with https agent options', () => {
             // Act
-            const connection = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
+            const got = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
                 agent: new https.Agent(),
             });
 
             // Assert
-            expect(connection).toBeTruthy();
-            expect(connection.options?.agent).toBeTruthy();
+            expect(got).toBeTruthy();
+            expect(got.options?.agent).toBeTruthy();
         });
     });
 });