+++ packages/axis-discovery-ssdp/test/network-interfaces/NetworkInterface.spec.ts
@@ -20,10 +20,10 @@ describe('NetworkInterface', () => {
             osMock.networkInterfaces.mockReturnValue(NETWORK_INTERFACE_WITH_TWO_ADDRESSES);
 
             // Act
-            const addresses = getIPv4Addresses();
+            const got = getIPv4Addresses();
 
             // Assert
-            expect(addresses).toEqual(['1.1.1.1', '2.2.2.2']);
+            expect(got).toEqual(['1.1.1.1', '2.2.2.2']);
         });
 
         test('should return addresses from multiple network interfaces', () => {
@@ -31,10 +31,10 @@ describe('NetworkInterface', () => {
             osMock.networkInterfaces.mockReturnValue(NETWORK_INTERFACES_WITH_TWO_ADDRESSES);
 
             // Act
-            const addresses = getIPv4Addresses();
+            const got = getIPv4Addresses();
 
             // Assert
-            expect(addresses).toEqual(['1.1.1.1', '2.2.2.2']);
+            expect(got).toEqual(['1.1.1.1', '2.2.2.2']);
         });
 
         test('should not return internal addresses', () => {
@@ -42,10 +42,10 @@ describe('NetworkInterface', () => {
             osMock.networkInterfaces.mockReturnValue(NETWORK_INTERFACES_WITH_INTERNAL_ADDRESSES);
 
             // Act
-            const addresses = getIPv4Addresses();
+            const got = getIPv4Addresses();
 
             // Assert
-            expect(Object.keys(addresses)).toHaveLength(0);
+            expect(Object.keys(got)).toHaveLength(0);
         });
 
         test('should not return IPv6 addresses', () => {
@@ -53,10 +53,10 @@ describe('NetworkInterface', () => {
             osMock.networkInterfaces.mockReturnValue(NETWORK_INTERFACES_WITH_IPV6_ADDRESSES);
 
             // Act
-            const addresses = getIPv4Addresses();
+            const got = getIPv4Addresses();
 
             // Assert
-            expect(Object.keys(addresses)).toHaveLength(0);
+            expect(Object.keys(got)).toHaveLength(0);
         });
 
         test('should not fail on systems without network interfaces', () => {
@@ -64,10 +64,10 @@ describe('NetworkInterface', () => {
             osMock.networkInterfaces.mockReturnValue(NO_NETWORK_INTERFACES);
 
             // Act
-            const addresses = getIPv4Addresses();
+            const got = getIPv4Addresses();
 
             // Assert
-            expect(Object.keys(addresses)).toHaveLength(0);
+            expect(Object.keys(got)).toHaveLength(0);
         });
     });
 });