+++ packages/axis-maintenance/test/shared/Connection.spec.ts
@@ -7,49 +7,49 @@ describe('connection', () => {
         describe('#ctor(protocol, ...)', () => {
             test('should return connection without options', () => {
                 // Act
-                const connection = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass');
+                const got = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass');
 
                 // Assert
-                expect(connection).toBeTruthy();
+                expect(got).toBeTruthy();
             });
 
             test('should return connection with http agent options', () => {
                 // Act
-                const connection = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
+                const got = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
                     agent: new http.Agent(),
                 });
 
                 // Assert
-                expect(connection).toBeTruthy();
-                expect(connection.options?.agent).toBeTruthy();
+                expect(got).toBeTruthy();
+                expect(got.options?.agent).toBeTruthy();
             });
 
             test('should return connection with https agent options', () => {
                 // Act
-                const connection = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
+                const got = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
                     agent: new https.Agent(),
                 });
 
                 // Assert
-                expect(connection).toBeTruthy();
-                expect(connection.options?.agent).toBeTruthy();
+                expect(got).toBeTruthy();
+                expect(got.options?.agent).toBeTruthy();
             });
         });
 
         test('should return URL given HTTP protocol', () => {
             // Act
-            const connection = new Connection(Protocol.Http, '1.2.3.4', 5678, 'root', 'pass');
+            const got = new Connection(Protocol.Http, '1.2.3.4', 5678, 'root', 'pass');
 
             // Assert
-            expect(connection.url).toBe('http://1.2.3.4:5678');
+            expect(got.url).toBe('http://1.2.3.4:5678');
         });
 
         test('should return URL given HTTPS protocol', () => {
             // Act
-            const connection = new Connection(Protocol.Https, '1.2.3.4', 5678, 'root', 'pass');
+            const got = new Connection(Protocol.Https, '1.2.3.4', 5678, 'root', 'pass');
 
             // Assert
-            expect(connection.url).toBe('https://1.2.3.4:5678');
+            expect(got.url).toBe('https://1.2.3.4:5678');
         });
     });
 });