+++ packages/axis-snapshot/src/Protocol.ts
@@ -2,12 +2,12 @@
  * Enum describing communication protocols.
  */
 export enum Protocol {
-  /**
-   * Hypertext Transfer Protocol (HTTP).
-   */
-  Http,
-  /**
-   * HTTP over Transport Layer Security (TLS).
-   */
-  Https,
+    /**
+     * Hypertext Transfer Protocol (HTTP).
+     */
+    Http,
+    /**
+     * HTTP over Transport Layer Security (TLS).
+     */
+    Https,
 }