+++ packages/axis-snapshot/test/Connection.spec.ts
@@ -1,35 +1,35 @@
-import { Agent as HttpAgent } from 'http';
-import { Agent as HttpsAgent } from 'https';
+import * as http from 'http';
+import * as https from 'https';
 import { Connection, Protocol } from '../src';
 
 describe('connection', () => {
     describe('#ctor(protocol, ...)', () => {
         test('should return connection without options', () => {
             // Act
-            const connection = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass');
+            const got = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass');
 
             // Assert
-            expect(connection.options?.agent).toBeUndefined();
+            expect(got.options?.agent).toBeUndefined();
         });
 
         test('should return connection with http agent options', () => {
             // Act
-            const connection = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
-                agent: new HttpAgent(),
+            const got = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
+                agent: new http.Agent(),
             });
 
             // Assert
-            expect(connection.options?.agent).toBeDefined();
+            expect(got.options?.agent).toBeDefined();
         });
 
         test('should return connection with https agent options', () => {
             // Act
-            const connection = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
-                agent: new HttpsAgent(),
+            const got = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass', {
+                agent: new https.Agent(),
             });
 
             // Assert
-            expect(connection.options?.agent).toBeDefined();
+            expect(got.options?.agent).toBeDefined();
         });
     });
 });