+++ packages/axis-maintenance/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### :policeman: Security
+
+- Replaced deprecated dependency [request](https://github.com/request/request) with [got](https://github.com/sindresorhus/got)
+
+### :dizzy: Changed
+
+- [BREAKING CHANGE] Removed properties `statusCode`, `cause` and `response` from `RequestError`
+- [BREAKING CHANGE] Renamed `UnauthorizationError` to `UnauthorizedError`
+
 ## [1.1.1] - 2020-05-12
 
 ### :syringe: Fixed