+++ packages/axis-snapshot/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### :policeman: Security
+
+- Replaced deprecated dependency [request](https://github.com/request/request) with [got](https://github.com/sindresorhus/got)
+
+### :dizzy: Changed
+
+- [BREAKING CHANGE] Removed properties `statusCode`, `cause` and `response` from `RequestError`
+
 ## [1.0.0] - 2020-10-22
 
 ### :zap: Added